refactor(notebook): reorganize components and update classnames
I think it is clearer to organize each component it is own directory
instead of putting all the files in a Notebook directory.
This way, I can split the CSS for each component and make it easier to
maintain.
I also remove the div wrapper for the notebook since all the things in
Main is Notebook related. So this node was useless.